requirements.txt
fastapi
uvicorn
azure-cognitiveservices-speech
